Moved API routes from __init__.py to routes.py.
