+++ pma_api/__init__.py
@@ -22,7 +22,7 @@ def create_app(config_name):
 
     db.init_app(app)
 
-    from .api_1_0 import api as api_1_0_blueprint
+    from .api_1_0.routes import api as api_1_0_blueprint
     app.register_blueprint(api_1_0_blueprint)
 
     return app
